ca: Profile tweaks (default, promote CA flag)
This adds the ability to flag a profile to promote the first domain/IP
to the common name. This was previously removed when promotion was
deprecated, but is still allowed in the Let's Encrypt "classic" profile,
so this helps mock this behavior (and also allows it to be mocked in CAs
that still do the same).